matchmakersite/jobs/requirements.txt
adium-theme-ubuntu==0.3.4
Django==1.10.2
django-localflavor==1.3
dotenv==0.0.5
importer==1.2
keen==0.3.26
Pillow==3.4.1
pycryptodome==3.4
pytz==2016.7
requests==2.10.0
unity-lens-photos==1.0
xia==1.1
